% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{access_sparse}
\alias{access_sparse}
\alias{sparse_indptr}
\alias{sparse_indices}
\alias{sparse_data}
\title{Access compressed sparse column format (CSC) from a matrix}
\usage{
access_sparse(M, slotName)

sparse_indptr(M)

sparse_indices(M)

sparse_data(M)
}
\arguments{
\item{M}{A \link{matrix}, or matrix-like coercible to \link[Matrix:CsparseMatrix-class]{Matrix::CsparseMatrix}.}

\item{slotName}{The name (\link{chracter}) of the CsparseMatrix slot.}
}
\value{
The values corresponding to the indicated compressed sparse column feature.
}
\description{
This relies heavily on the compressed sparse format
defined in the \link[Matrix:Matrix]{Matrix::Matrix} package \link[Matrix:CsparseMatrix-class]{Matrix::CsparseMatrix}.
}
\details{
See a
\href{http://www.scipy-lectures.org/advanced/scipy_sparse/csc_matrix.html}{python doc on CSC}
}
\section{Functions}{
\itemize{
\item \code{sparse_indptr}: Access the index pointer, indptr

\item \code{sparse_indices}: Access the indices

\item \code{sparse_data}: Access the sparse data entries
}}

\examples{
mat1 = matrix(c(1,0,2,0,0,3,4,5,6), nrow = 3, byrow = TRUE)
access_sparse(mat1, "p")
sparse_indptr(mat1)
sparse_indices(mat1)
sparse_data(mat1)
}
