% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{add_biom_h5_attributes}
\alias{add_biom_h5_attributes}
\title{Add top-level attributes to a biom-file v2.1}
\usage{
add_biom_h5_attributes(biomList, biomFile)
}
\arguments{
\item{biomList}{A list with structure/names that match the
biom-format version 2.1 structure.
For example, the output of \code{\link[=matrix_to_biom_list_2.1]{matrix_to_biom_list_2.1()}}.}

\item{biomFile}{An open and writable object of the \link[=H5File-class]{hdf5r}
(e.g. \code{H5File$new(File, mode="w")}).
Regular users of biomformat package should not define this directly,
but instead get it via constructor methods that also include
valid attributes and groups, e.g. \code{\link[=write_biom_list_2.1]{write_biom_list_2.1()}}.}
}
\description{
Add top-level attributes to a biom-file v2.1
}
