% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{biom_header_2.1}
\alias{biom_header_2.1}
\title{Generate required top-level attributes for biom version 2.1}
\usage{
biom_header_2.1(id = "", type = biomTypes2.1,
  format_url = "http://biom-format.org", format_version = c(2L, 1L),
  generated_by = sprintf("biomformat \%s", packageVersion("biomformat")),
  creation_date = strptime(x = Sys.time(), format =
  "\%Y-\%m-\%d \%H:\%M:\%S"))
}
\arguments{
\item{id}{Optional table identifier string.}

\item{type}{String indicating the table type, from a controlled vocabulary, [`biomTypes2.1`].}

\item{format_url}{The URL string for the biom-format project.}

\item{format_version}{The biom-format version tuple.}

\item{generated_by}{The package and version that generated the corresponding biom file.}

\item{creation_date}{The date the file was built, in ISO 8601 format.}
}
\description{
See [the format definition](http://biom-format.org/documentation/format_versions/biom-2.1.html).
Note that this excludes the two matrix-dependent top-level attributes,
`shape` and `nnz`.
}
