% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPackage.R
\docType{package}
\name{biomformat-package}
\alias{biomformat-package}
\alias{biomformat}
\title{biomformat: An R interface to the BIOM file format}
\description{
This is an R package for interfacing with the biom-format.
It includes basic utilities for reading and writing BIOM format using R,
and native R methods for interacting with biom-data
that maps to functionality in other languages
that support biom-format, like python.
}
\references{
\url{http://www.biom-format.org/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joey711/biomformat/}
  \item \url{http://biom-format.org/}
  \item Report bugs at \url{https://github.com/joey711/biomformat/issues}
}

}
\author{
Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\keyword{package}
