% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\docType{data}
\name{biomTypes2.1}
\alias{biomTypes2.1}
\title{Acceptable values for biom v2.1 top-level \code{type}}
\format{An object of class \code{character} of length 7.}
\usage{
biomTypes2.1
}
\description{
Acceptable values for biom v2.1 top-level \code{type}
}
\keyword{datasets}
