% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-methods.R
\name{h5writeAttribute.scalar}
\alias{h5writeAttribute.scalar}
\title{write string variables to scalar hdf5 space}
\usage{
\method{h5writeAttribute}{scalar}(attr, h5obj, name)
}
\arguments{
\item{attr}{Character string.  Attribute to be saved.}

\item{h5obj}{An object of class H5IdComponent representing a H5 object identifier (file, group, or dataset). See H5Fcreate, H5Fopen, H5Gcreate, H5Gopen, H5Dcreate, or H5Dopen to create an object of this kind.}

\item{name}{Character string. Name of attribute.}
}
\description{
write string variables to scalar hdf5 space
}
