% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{matrix_to_biom_2.1_list}
\alias{matrix_to_biom_2.1_list}
\title{Convert matrix to biom-format 2.1 as an R list}
\usage{
matrix_to_biom_2.1_list(M, ...)
}
\arguments{
\item{M}{A \link{matrix}, or matrix-like coercible to \link[Matrix:CsparseMatrix-class]{Matrix::CsparseMatrix}.}

\item{...}{Additional named arguments passed to \code{\link[=biom_header_2.1]{biom_header_2.1()}}.}
}
\description{
See \href{http://biom-format.org/documentation/format_versions/biom-2.1.html}{The version 2.1 specification}
for more details.
}
\seealso{
\code{\link[=biom_header_2.1]{biom_header_2.1()}}

\code{\link[=matrix_to_biom_matlist]{matrix_to_biom_matlist()}}
}
