% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{matrix_to_biom_csc}
\alias{matrix_to_biom_csc}
\title{Transform matrix to biom-format list representation of Compressed Sparse Column matrix}
\usage{
matrix_to_biom_csc(m)
}
\arguments{
\item{m}{A base \link{matrix}.}
}
\description{
Transform matrix to biom-format list representation of Compressed Sparse Column matrix
}
\seealso{
A \href{http://www.scipy-lectures.org/advanced/scipy_sparse/csc_matrix.html}{python doc on CSC}
}
