% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{matrix_to_biom_matlist}
\alias{matrix_to_biom_matlist}
\title{Transform matrix to biom format (sparse) matrix list}
\usage{
matrix_to_biom_matlist(m)
}
\arguments{
\item{m}{The matrix that you want to coerce to a biom-like list.}
}
\description{
Assumes that rows are observations and columns are samples.
}
