% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-methods.R
\name{read_hdf5_biom}
\alias{read_hdf5_biom}
\title{Read in a biom-format v2 (HDF5) file, returning a \code{list}.}
\usage{
read_hdf5_biom(biom_file)
}
\arguments{
\item{biom_file}{(Required). A character string path to an HDF5-format
BIOM file.}
}
\value{
A named list representing the BIOM data, suitable for passing to
 \code{\link{biom}()}.
}
\description{
This function reads a BIOM v2 HDF5 file directly. In normal use, you should
call \code{\link{read_biom}}, which automatically routes to this function
when it detects the HDF5 magic bytes. Call this function directly only when
you are certain the file is HDF5 format.
}
\details{
The \code{rhdf5} package is required at runtime. If it is not available, or
if the underlying HDF5 system libraries are absent, a graceful R-level
warning is thrown instead of a fatal C-level abort.
}
\examples{
biom_file <- system.file("extdata", "rich_sparse_otu_table_hdf5.biom", package = "biomformat")
x = read_hdf5_biom(biom_file)
x = biom(x)
outfile = tempfile()
write_biom(x, outfile)
y = read_biom(outfile)
identical(observation_metadata(x),observation_metadata(y))
identical(sample_metadata(x),sample_metadata(y))
identical(biom_data(x), biom_data(y))
}
\references{
\url{http://biom-format.org/}
}
\seealso{
\code{\link{read_biom}}, \code{\link{make_biom}},
\code{\link{biom-class}}, \code{\link{write_biom}},
\code{\link{write_hdf5_biom}}.
}
