% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\docType{data}
\name{requiredAttributes2.1}
\alias{requiredAttributes2.1}
\title{The top-level attributes required according to biom-format 2.1 specification}
\format{An object of class \code{character} of length 8.}
\usage{
requiredAttributes2.1
}
\description{
See \href{http://biom-format.org/documentation/format_versions/biom-2.1.html}{the format definition}.
}
\keyword{datasets}
