% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-methods.R
\name{write_biom_hdf5}
\alias{write_biom_hdf5}
\title{Write biom file in HDF5 format}
\usage{
write_biom_hdf5(biom, file)
}
\arguments{
\item{biom}{A biom object.}

\item{file}{A character string indicating the 
file location of the biom formatted file. This is a HDF5 formatted file
specific to biological datasets. 
The format is formally defined at 
\href{http://biom-format.org/documentation/biom_format.html}{the biom-format definition}}
}
\description{
Write biom file in HDF5 format
}
