% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{write_biom_list_2.1}
\alias{write_biom_list_2.1}
\title{Write biom-format version 2.1 to HDF5 file}
\usage{
write_biom_list_2.1(biomList, File = "newFile.biom")
}
\arguments{
\item{biomList}{A list with structure/names that match the
biom-format version 2.1 structure.
For example, the output of \code{\link[=matrix_to_biom_list_2.1]{matrix_to_biom_list_2.1()}}.}

\item{File}{A character string indicating the destination file}
}
\value{
Invisible returns an object of the \linkS4class{H5File},
after having used the same object to write to \code{File}.
Note that the corresponding file will have been closed
via the \code{$close_all()} method.
}
\description{
Write biom-format version 2.1 to HDF5 file
}
