% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-methods.R
\name{as_tibble.biom}
\alias{as_tibble.biom}
\title{Coerce a biom object to a tidy tibble}
\usage{
as_tibble.biom(x, ...)
}
\arguments{
\item{x}{A \code{\link{biom-class}} object.}

\item{...}{Passed to \code{\link[tibble]{as_tibble}}.}
}
\value{
A \code{\link[tibble]{tibble}} in long (tidy) format.
  See \code{\link{as.data.frame.biom}} for column details.
}
\description{
A thin wrapper around \code{\link{as.data.frame.biom}} that returns a
\code{\link[tibble]{tibble}} instead of a plain \code{data.frame}.
Requires the \pkg{tibble} package (in \code{Suggests}).
}
\examples{
biom_file <- system.file("extdata", "rich_dense_otu_table.biom",
                         package = "biomformat")
x <- read_biom(biom_file)
if (requireNamespace("tibble", quietly = TRUE)) {
  # Call the method directly to ensure correct S3 dispatch
  tbl <- as_tibble.biom(x)
  tbl
}
}
\seealso{
\code{\link{as.data.frame.biom}}
}
